Agents_Library/Agents_Collection/requirements.txt
# beautifulsoup4==4.13.4
# requests==2.32.4
# python-dateutil==2.9.0.post0
# langchain==0.3.25
# langchain-community==0.3.25
# langchain-google-genai==2.1.5
# langchain-experimental==0.3.4
# langchain-neo4j==0.4.0
# langchain-pinecone==0.2.8
# python-dotenv==1.0.1
# openai==1.90.0
# streamlit==1.32.2
# fastapi==0.115.13
# uvicorn==0.34.3
# httpx==0.28.1
# pydantic==2.11.7
# pinecone-client==3.0.0
# pyreadline3==3.5.4
# pypdf==5.9.0
# python-multipart==0.0.20
# PyPDF2==3.0.1
# names==0.3.0
# pyngrok==7.3.0
# langdetect==1.0.9
# email-validator 
# tenacity
# psycopg2-binary
# supabase

# Compare this snippet from Agents-Collection/Ashutosh-Knowledge-Base-Agent/requirements.txt:
# jinja2==3.1.4
# starlette==0.38.5

# fastapi==0.114.2
# uvicorn[standard]==0.30.6
# python-dotenv==1.0.1
# openai==1.46.0
# pinecone-client==3.2.2
# jinja2==3.1.4
# starlette==0.38.5
# pydantic==2.9.2
# httpx==0.27.2


###########################################################


beautifulsoup4
requests
python-dateutil
langchain
langchain-community
langchain-google-genai
langchain-experimental
langchain-neo4j
langchain-pinecone
python-dotenv
openai
streamlit
fastapi
uvicorn
httpx
pydantic
pinecone-client
pyreadline3
pypdf
python-multipart
PyPDF2
names
pyngrok
langdetect
email-validator
tenacity
psycopg2-binary
jinja2
starlette
pandas 
openpyxl
tqdm
langgraph
tabulate
matplotlib 
Seaborn 